requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0
redis==5.0.1
dj-database-url==2.1.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
pandas==2.2.0
openpyxl==3.1.2
reportlab==4.0.8
social-auth-app-django==5.4.0
